Only send ProcessPayment request once
Updating the useEffect to take an empty array as it's second arg, so the
useEffect callback only gets called once when the PaymentProcessor
component is mounted. Because the QR callback is executed every time the
QR barcode is moved in the camera frame this caused the PaymentProcess
component to be re-rendered multiple times, causing the network request
to be sent multiple times.