Reverting to Promise chain from async/await
Because async functions always return a promise it can't be used
directly as the useEffect function, as useEffect either needs to return
nothing or a function that can be called on cleanup. It's easier to just
revert to the promise chain syntax, than mess around with nested
functions or IIFE at this point. Also added an additional link in the
chain to parse the response as JSON.