Add PaymentCOnfirmation page
Refactoring the payment confirmation out of the Payment Processor. I'd
prefer to use more native tooling to present options to the user instead
of the system alert box, as it's more inline with what a more realistic
scenario would do. Additionally, it helps to break up the request flow
between components and makes it easier to read.