Only scan the QR code if scanner is active
Because react-navigation keeps all screens mounted the QR scanner
component is still active, so if the QR code was re-scanned during the
rest of the payment flow the app would be re-navigated to the
PaymentProcessing screen and get stuck. Using the useIsFocus hook to
only process the QR code if the QR scanner component is currently
focused on.