Redirect to app link when mobile landing URL is visited
the mobile link is configured to point to /mobile/landing on the
provider, so the provider is responsible for redirecting to the app
link, so the url opens in the KomojuDemo app. The session query
parameter is being URI encoded as it's is being forwarded to the mobile
app as a sub resource, instead of a query parameter. I couldn't get the
query parameter to be supported by the React Native app so URI encoding
the address and decoding it in the app was the next best solution.