Add script to build tke APK release
Simple script to build the APK file and move it to the top level of the
repository. This seemed easier then messing with the gitignore rules to
get the APK file tracked while still ignoring everything else, and means
people don't have to dig through the repository file structure to find
it.